requirements.txt
numpy
trimesh
matplotlib
scikit-learn
tqdm
markdown
reportlab
Pillow
